Refactor: Updating modularity for new pages
